Improve Combat player death information
Improve Combat player death information:
 - reduce wait time till information load
 - reduce memory use
 - improve web UI view

+++ src/CombatParser/CombatParser.Domain/Entities/CombatPlayerData/CombatPlayerDeath.cs
@@ -4,7 +4,7 @@ public record CombatPlayerDeath(
     TimeSpan Time,
     string Name,
     string Spell,
-    int Value,
+    long Value,
     long CurrentHealth,
     long MaxHealth,
     int Status,
